+++ e2e/test-env/src/client/RicosTestApp/RicosTestApp.jsx
@@ -9,6 +9,7 @@ import { default as viewerPlugins } from './viewerPlugins';
 import './styles.global.scss';
 import theme from '../../../../../examples/main/shared/theme/theme';
 import { testVideos } from '../../../../../examples/main/shared/utils/mock';
+import { previewSettings } from 'wix-rich-content-preview';
 import { TextSelectionToolbar, TwitterButton } from 'wix-rich-content-text-selection-toolbar';
 
 const onVideoSelected = (url, updateEntity) => {
@@ -63,6 +64,7 @@ class RicosTestApp extends PureComponent {
           locale={locale}
           cssOverride={theme}
           seoSettings={seoMode}
+          preview={testAppConfig.showDefaultPreview && previewSettings()}
         />
         <TextSelectionToolbar container={this.viewerRef.current}>
           {selectedText => <TwitterButton selectedText={selectedText} />}