+++ packages/preview/web/src/Components/default-transformation.js
@@ -1,5 +1,35 @@
 import ContentStateTransformation from '../RuleEngine/ContentStateTransformation';
 
+const galleryStyle = {
+  galleryLayout: 2,
+  gallerySizeType: 'px',
+  gallerySizePx: 300,
+  galleryMargin: 0,
+  oneRow: false,
+  cubeRatio: 1,
+  galleryThumbnailsAlignment: 'bottom',
+  isVertical: false,
+  imageMargin: 20,
+  thumbnailSpacings: 0,
+  cubeType: 'fill',
+  enableInfiniteScroll: true,
+  titlePlacement: 'SHOW_ON_HOVER',
+  allowHover: false,
+  itemClick: 'link',
+  fullscreen: false,
+  showArrows: false,
+  gridStyle: 1,
+  loveButton: false,
+  allowSocial: false,
+  allowDownload: false,
+  mobileSwipeAnimation: 'NO_EFFECT',
+  thumbnailSize: 120,
+  gotStyleParams: true,
+  showVideoPlayButton: false,
+  videoPlay: 'auto',
+  numberOfImagesPerRow: 2,
+};
+
 export const defaultTransformation = new ContentStateTransformation({
   _if: metadata => metadata.allText.length > 0,
   _then: (metadata, preview) => {
@@ -7,50 +37,34 @@ export const defaultTransformation = new ContentStateTransformation({
   },
 })
   .rule({
-    _if: metadata => metadata.galleryItems.length > 0 && metadata.galleryItems.length < 5,
+    _if: metadata => metadata.galleryItems.length === 1,
     _then: (metadata, preview) => {
       const mediaInfo = metadata.galleryItems[0];
       const type = mediaInfo.type;
-      return preview[type]({ mediaInfo }).seeFullPost();
+      return preview[type]({ mediaInfo });
     },
   })
   .rule({
-    _if: metadata => metadata.galleryItems.length >= 5,
+    _if: metadata => metadata.galleryItems.length > 1 && metadata.galleryItems.length <= 4,
     _then: (metadata, preview) =>
       preview
         .gallery({
           mediaInfo: metadata.galleryItems.slice(0, 4),
           overrides: {
-            styles: {
-              galleryLayout: 2,
-              gallerySizeType: 'px',
-              gallerySizePx: 300,
-              galleryMargin: 0,
-              oneRow: false,
-              cubeRatio: 1,
-              galleryThumbnailsAlignment: 'bottom',
-              isVertical: false,
-              imageMargin: 20,
-              thumbnailSpacings: 0,
-              cubeType: 'fill',
-              enableInfiniteScroll: true,
-              titlePlacement: 'SHOW_ON_HOVER',
-              allowHover: false,
-              itemClick: 'link',
-              fullscreen: false,
-              showArrows: false,
-              gridStyle: 1,
-              loveButton: false,
-              allowSocial: false,
-              allowDownload: false,
-              mobileSwipeAnimation: 'NO_EFFECT',
-              thumbnailSize: 120,
-              gotStyleParams: true,
-              showVideoPlayButton: false,
-              videoPlay: 'auto',
-              numberOfImagesPerRow: 2,
-            },
+            styles: galleryStyle,
           },
         })
-        .imageCounter({ counter: metadata.images.length - 4 }),
+        .seeFullPost(),
+  })
+  .rule({
+    _if: metadata => metadata.galleryItems.length > 4,
+    _then: (metadata, preview) =>
+      preview
+        .gallery({
+          mediaInfo: metadata.galleryItems.slice(0, 4),
+          overrides: {
+            styles: galleryStyle,
+          },
+        })
+        .imageCounter({ counter: metadata.galleryItems.length - 4 }),
   });